move all vercel-edge functions into fern-docs-edge package
