+++ packages/ui/docs-bundle/src/pages/api/fern-docs/api-definition/[api]/webhook/[webhook].ts
@@ -1,8 +1,8 @@
 import { ApiDefinitionLoader } from "@/server/ApiDefinitionLoader";
 import { getXFernHostNode } from "@/server/xfernhost/node";
 import * as ApiDefinition from "@fern-api/fdr-sdk/api-definition";
+import { getFeatureFlags } from "@fern-ui/fern-docs-edge";
 import { NextApiHandler, NextApiResponse } from "next";
-import { getFeatureFlags } from "../../../feature-flags";
 
 const resolveApiHandler: NextApiHandler = async (req, res: NextApiResponse<ApiDefinition.ApiDefinition>) => {
     const xFernHost = getXFernHostNode(req);
