+++ packages/ui/docs-bundle/src/server/auth/getApiKeyInjectionConfig.ts
@@ -1,8 +1,8 @@
+import { getAuthEdgeConfig } from "@fern-ui/fern-docs-edge";
 import type { APIKeyInjectionConfig } from "@fern-ui/ui/auth";
 import type { NextApiRequestCookies } from "next/dist/server/api-utils";
 import type { NextRequest } from "next/server";
 import { OAuth2Client } from "./OAuth2Client";
-import { getAuthEdgeConfig } from "./getAuthEdgeConfig";
 
 // TODO: since this is for ORY (rightbrain) only, lets refactor
 export async function getAPIKeyInjectionConfig(
