+++ packages/ui/fern-docs-edge/.prettierrc.cjs
@@ -1,0 +1,1 @@
+module.exports = require("../../../.prettierrc.json");
