+++ packages/ui/fern-docs-edge/.stylelintrc.json
@@ -1,0 +1,1 @@
+{ "extends": ["../../../shared/stylelintrc.shared.json"] }
