+++ packages/ui/fern-docs-edge/src/getSeoDisabled.ts
@@ -1,5 +1,5 @@
+import { isCustomDomain } from "@fern-ui/fern-docs-utils";
 import { get } from "@vercel/edge-config";
-import { isCustomDomain } from "./isCustomDomain";
 
 export async function getSeoDisabled(host: string): Promise<boolean> {
     if (!isCustomDomain(host)) {
