+++ packages/ui/fern-docs-edge/tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "extends": "@fern-platform/configs/tsconfig/library.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  },
+  "include": ["./src/**/*"],
+  "references": [
+    {
+      "path": "../fern-docs-utils"
+    },
+    {
+      "path": "../app"
+    }
+  ]
+}
