+++ packages/ui/fern-docs-utils/src/constants.ts
@@ -4,6 +4,7 @@ export const COOKIE_FERN_DOCS_PREVIEW = "_fern_docs_preview" as const;
 export const COOKIE_FERN_TOKEN = "fern_token" as const;
 export const COOKIE_ACCESS_TOKEN = "access_token" as const; // for api key injection
 export const COOKIE_REFRESH_TOKEN = "refresh_token" as const; // for api key injection
+export const COOKIE_EMAIL = "email" as const; // for api key injection
 export const HEADER_X_FERN_HOST = "x-fern-host" as const;
 
 /**
