+++ packages/ui/fern-docs-utils/src/isCustomDomain.ts
